GRIDEDIT-324: Correct handling of curvilinear grids with invalid nodes requires specifying a range composed of only valid nodes in the root finding algorithm. + renamed ComputeClosestPointOnSpline to ComputeClosestPointOnSplineSegment + bug fixing in ComputeInvalidHorizontalBoundaryNodes, ComputeInvalidVerticalBoundaryNodes, ProjectOnHorizontalBoundaries (rightNode instead of leftnode) + added orthogonalization unit test on a curvi grid with missing faces

+++ include/MeshKernel/Operations.hpp
@@ -419,7 +419,7 @@ namespace meshkernel
 
         const double eps = 1e-5;
         const double splFac = 1.0;
-        const auto intCoordinate = static_cast<size_t>(std::floor(pointAdimensionalCoordinate));
+        const auto intCoordinate = static_cast<double>(std::floor(pointAdimensionalCoordinate));
         if (pointAdimensionalCoordinate - intCoordinate < eps)
         {
             return pointCoordinate = coordinates[intCoordinate];
